deps/v8/test/webkit/string-sort-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This page tests sorting of string values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "12" < "123" is true
PASS ["12","123"].sort().toString() is "12,123"
PASS "123" < "13" is true
PASS ["123","13"].sort().toString() is "123,13"
PASS "" < " " is true
PASS [""," "].sort().toString() is ", "
PASS "close" < "closed" is true
PASS ["close","closed"].sort().toString() is "close,closed"
PASS "a" < "~" is true
PASS ["a","~"].sort().toString() is "a,~"
PASS "_" < "a" is true
PASS ["_","a"].sort().toString() is "_,a"
PASS "0" < "A" is true
PASS ["0","A"].sort().toString() is "0,A"
PASS "-" < "0" is true
PASS ["-","0"].sort().toString() is "-,0"
PASS " " < "-" is true
PASS [" ","-"].sort().toString() is " ,-"
PASS "A" < "ABC" is true
PASS ["A","ABC"].sort().toString() is "A,ABC"
PASS successfullyParsed is true

TEST COMPLETE
